Moving redis lock into redis module
